Move www_test related files out of the main lib
