+++ src/lib.rs
@@ -1,12 +1,6 @@
-#![feature(impl_trait_in_assoc_type)]
 #![no_std]
 
-pub mod board;
-pub mod config;
-pub mod html;
-pub mod power;
-pub mod server;
-pub mod simple_output;
-pub mod wifi;
+mod board;
+mod power;
 
 pub use board::Board;
